change the type of counter values from int to u64

+++ hash.h
@@ -4,8 +4,8 @@
 #include "memory_access.h"
 
 void* create_hash(void);
-void add_to_hash(void* _hash, u64 key, int value);
-int get_from_hash(void* _hash, u64 key);
+void add_to_hash(void* _hash, u64 key, u64 value);
+u64 get_from_hash(void* _hash, u64 key);
 int get_size_of_hash(void* _hash);
 struct memory_access* get_memory_access(void* _hash);
